Create initialState in src/index.js
